add sampleTest of list to dataFrame. TODO: Add read and write csv

+++ app/src/main/java/com/arnold/sleepmonitor/data_structure/SingleTimeData.kt
@@ -7,5 +7,5 @@ data class SingleTimeData(
     val acc_y: Double,
     val acc_z: Double,
     val volume: Double,
-    val frequency: Int
+    val frequency: Double
 )
